+++ cx-portal/src/components/overlays/AddUser/AddUserContent.tsx
@@ -34,7 +34,7 @@ import {
   usersToAddSelector,
 } from 'features/admin/userDeprecated/slice'
 import { useDispatch, useSelector } from 'react-redux'
-import { closeOverlay } from 'features/control/overlay/actions'
+import { closeOverlay } from 'features/control/overlay'
 import './AddUserOverlay.scss'
 import { SingleUserContent } from './SingleUserContent'
 import {