+++ cx-portal/src/components/overlays/AppOverViewConfirm/index.tsx
@@ -25,7 +25,7 @@ import {
   DialogHeader,
   Typography,
 } from 'cx-portal-shared-components'
-import { closeOverlay } from 'features/control/overlay/actions'
+import { closeOverlay } from 'features/control/overlay'
 import { useTranslation } from 'react-i18next'
 import { useDispatch } from 'react-redux'
 import './AppOverviewConfirm.scss'