+++ cx-portal/src/components/pages/Home/components/StageSection/index.tsx
@@ -22,6 +22,7 @@ import { Button, MainHeader } from 'cx-portal-shared-components'
 import { useTranslation } from 'react-i18next'
 import PageService from 'services/PageService'
 import { label as BusinessApplications } from '../BusinessApplicationsSection'
+import { getAssetBase } from 'services/EnvironmentService'
 
 export default function SearchSection() {
   const { t } = useTranslation()
@@ -34,7 +35,7 @@ export default function SearchSection() {
         headerHeight={731}
         subTitleWidth={787}
         background="LinearGradient1"
-        imagePath="./home-stage-desktop.png"
+        imagePath={`${getAssetBase()}/images/frame/home-stage-desktop.png`}
       >
         <Button
           sx={{ margin: '40px 10px 0 0' }}